Confirmation email and failure
Added examples to user_confirmations_spec for the confirmation email and
confirmation failure when a user is already confirmed.

Added signup_someone to utilities to create an easily accessible static
user that can be used in tests. Setting the user is as simple as
let(:user) { User.find_by_email('someone@example.com'). Now there is an
unconfirmed user to use in tests.

Customized the confirmations_controller with specific redirect and flash
message.

Added custom flash message to devise.en.yml.

Added a custom sessions controller and updated routes to use it.

All examples are green.